requirements.txt
asgiref==3.8.1
Django==5.1.2
django-filter==25.1
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
pillow==11.0.0
PyJWT==2.10.1
sqlparse==0.5.1
tzdata==2024.2
